+++ packages/eslint-plugin/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Enhancements
+
+-   Update `use-recommended-components` rule to prefer `@wordpress/ui` `Field` over legacy `@wordpress/components` `BaseControl` ([#82095](https://github.com/WordPress/gutenberg/pull/82095)).
+
 ## 25.10.0 (2026-08-26)
 
 ### Bug Fixes